pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bark-mcp"
version = "0.1.1"
description = "MCP server talking to Bark notification server for iOS"
readme = "README.md"
authors = [
    {name = "Bark MCP Contributors"}
]
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "fastmcp>=2.0.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
bark-mcp = "bark_mcp.cli:main"

[project.urls]
"Homepage" = "https://github.com/yourusername/bark-mcp"
"Bug Tracker" = "https://github.com/yourusername/bark-mcp/issues"
